+++ lib/types/model.type.ts
@@ -1,0 +1,8 @@
+export type ModelType = {
+  title: string;
+  position: {
+    x: number;
+    y: number;
+  };
+  image: File;
+};