+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+### [0.0.3] 2025.06.02
+
+- Added three js fiber
+- Imported test Cube models
+
 ### [0.0.2] 2025.06.02
 
 ## Added